Bug 1983028 - Part 2: Sync Utils-vixl with upstream. r=spidermonkey-reviewers,iain
This appears to fix some UB issues when negating unsigned values.

Differential Revision: https://phabricator.services.mozilla.com/D261100

UltraBlame original commit: 6f61a881b811a140deb0f697fb7f3256c319c2d4

+++ js/src/jit/arm64/vixl/Utils-vixl.cpp
@@ -391,7 +391,7 @@ float FPToFloat(double value,
   }
 
   VIXL_UNREACHABLE();
-  return value;
+  return static_cast<float>(value);
 }
